requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1
requests==2.32.3
langchain-core==0.2.14
langchain-community==0.2.8
langchain-text-splitters==0.2.1
langchain-google-genai==0.1.3
langchain-mcp-adapters==0.1.0
pypdf==4.2.0
pinecone-client==3.2.2
google-generativeai==0.6.6
duckduckgo-search==4.1.2
mcp==0.5.3
